# api/docs/openapi/admin.yaml
openapi: 3.0.3
info:
  title: EnrollAI - Admin API
  description: API for administrative functions in EnrollAI, allowing admins to manage users, courses, departments, and programs.
  version: 1.0.0
servers:
  - url: http://localhost:9000
    description: Local development environment
paths:
...
  /admin/courses:
    get:
      summary: Get all courses
      description: Retrieve a list of all courses.
      operationId: getAllCourses
      responses:
        '200':
          description: A list of courses retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
              example: [{"code": "CS101", "title": "Introduction to Computer Science", "description": "Basic concepts of computing.", "credits": 3}]
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "An unexpected error occurred."

    post:
      summary: Add a new course
      description: Create a new course.
      operationId: createCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Course'
            example:
              code: "CS101"
              title: "Introduction to Computer Science"
              description: "This course introduces the fundamentals of computer science."
              credits: 3
      responses:
        '201':
          description: Course created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Invalid input, object invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Invalid input data for course creation."
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "An unexpected error occurred."
components:
  schemas:
...
    Course:
      type: object
      properties:
        code:
          type: string
          description: Unique course code.
          example: "CS101"
        title:
          type: string
          description: Title of the course.
          example: "Introduction to Computer Science"
        description:
          type: string
          description: Detailed description of the course.
          example: "This course introduces the fundamentals of computer science."
        credits:
          type: integer
          format: int32
          description: Number of credits assigned to the course.
          example: 3
      required:
        - code
        - title
        - credits
    Error:
      type: object
      properties:
        message:
          type: string
          description: Detailed error message.
          example: "An unexpected error occurred."